Organize moe and mlp into variations
This is first step towards cleaning the model.py
